Move setup_legacy_cert_paths fixture to conftest for reuse
